% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin}
\alias{admin}
\alias{.setAdmin}
\alias{.getAdmin}
\title{Administration functions for smif.package}
\usage{
.setAdmin(x = TRUE)

.getAdmin()
}
\description{
Used by various functions to hide advanced functionality
}
\keyword{internal}
