% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHoldings.SMIF.R
\name{canConnect}
\alias{canConnect}
\title{Determines if connection is possible}
\usage{
canConnect(test.site = "8.8.8.8", n = 1, timeout = 1000, clean = TRUE)
}
\arguments{
\item{test.site}{Character; the site URL or IP address to test connection to.
Defaults to "8.8.8.8", Google's DNS-A server. Function}

\item{n}{Numeric; number of times to try the connection. Defaults to 1}

\item{timeout}{Numeric; time (in milliseconds) to try to connect before
returning \code{FALSE}. Defaults to 1000ms (1s)}

\item{clean}{Logical; should the input \code{test.site} be parsed automatically?
Defaults to \code{TRUE}}
}
\value{
Logical; logical vector corresponding to whether or not a connection
could be made to given site(s). Length will be equal to \code{length(test.site)}
}
\description{
Pings the specified URL(s) and/or IP address(es) to determine if connection
can be made/internet is up/websites are up.
}
\note{
To simply test if internet connection is possible, run canConnect() or
canConnect
To test specific sites: Google A: "8.8.8.8", Google B: "8.8.4.4",
Google Search: "www.google.com" or "google.com"
}
\examples{
canConnect()
}
