% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\name{cleanAccount}
\alias{cleanAccount}
\title{Clean accounting data}
\usage{
cleanAccount(acc)
}
\arguments{
\item{acc}{a character vector of a number in accounting format (i.e. "$792.76B")}
}
\value{
a numeric representation of the input string
}
\description{
Clean accounting data
}
\note{
Data with a suffix of a single "m" or "M" character will not be considered "1000"
but will be interpreted as one million (1000000).
}
