% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\name{cleanIP}
\alias{cleanIP}
\title{Cleans the name of a site}
\usage{
cleanIP(site.name)
}
\arguments{
\item{site.name}{Character; a string or vector of strings of the URL(s)
or IP address(es) to be cleaned}
}
\value{
Character; the processed URL(s)/IP address(es)
}
\description{
Attempts to coerce a full or partial URL into a form that can be used
by calls to system functions. A small number of keyword inputs have
special cases. Some 'notable' sites have IP addresses specified and
return those instead of a processed URL.
}
\note{
Due to the fact that some data firms employ multiple IP addresses
for different purposes, the IP addresses returned for 'notable' sites
are not guarenteed to correspond to the IP addresses accessed by many
popular R functions, such as those in \code{quantmod}, though most are.
}
\examples{
cleanIP("http://www.stevens.edu")
}
\seealso{
Other data cleaning functions: \code{\link{cleanAccount}},
  \code{\link{cleanIndex}}, \code{\link{cleanSector}}
}
\keyword{internal}
