% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\name{cleanSector}
\alias{cleanSector}
\title{Clean sector names}
\usage{
cleanSector(sector, verbose = getOption("verbose", FALSE))
}
\arguments{
\item{sector}{Character; string to be transformed into a uniform sector}

\item{verbose}{Logical; whether the function should throw messages. Defaults to getOption("verbose")}
}
\value{
Character; string representation of a sector
}
\description{
This function takes a character vector input and parses through the SMIF Asset Allocation
\code{sectors} dataset and attempts to coerce the input into a sector compliant with SMIF AA
sector formatting guidelines.
}
\examples{
cleanSector("Consumer Non-Durables")
}
\seealso{
Other data cleaning functions: \code{\link{cleanAccount}},
  \code{\link{cleanIndex}}
}
\concept{
clean index
}
