% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{getAV}
\alias{getAV}
\title{Get SMIF AlphaVantage API key}
\usage{
getAV()
}
\value{
Character; the API key as a string
}
\description{
Retrieves the SMIF's API key for Alpha Vantage, allowing data fetching
using \code{quantmod}'s \code{getSymbols.av}. This API key is used inside
\code{\link{getSymbols.SMIF}} when fetching data from AlphaVantage is
both enabled, and possible.
}
\examples{
getAV()
}
\keyword{internal}
