% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHoldings.SMIF.R
\name{getHoldings.SMIF}
\alias{getHoldings.SMIF}
\title{Loads the current SMIF holdings}
\usage{
getHoldings.SMIF(adv = F, ...)
}
\arguments{
\item{adv}{Logical; should the data be returned unformatted. Defaults to \code{FALSE}}

\item{...}{additional parameters}
}
\value{
a data.frame with the following columns:
\item{Ticker}{The ticker of the given holding}
\item{Shares}{The number of shares we currently own}
}
\description{
Fetches the current holdings for the SMIF from the server. Data returned will be sufficient to conduct
portfolio optimization on its own. Additional analysis on the information is possible in conjunction with
the other \code{smif.package} functions.
}
\details{
For compliance purposes, in order to connect to the server the connection info is not stored in plaintext.
The password is inputted directly by the user at runtime, and the server's IP address is encrypted. Decryption
occurs at runtime via answering a question that members are aware of but the general public does not know.
Function uses specific error handling and custom error messages.
}
\section{Warning}{

Due to security settings on our SQL server, for certain R sessions the DBI connections are severed improperly.
When DBI connections are left open too long they expire and subsequent DBI connections to the server are
impossible until a new session is started. For this reason it is reccomended the results be stored in an
environment variable as soon as they are downloaded, and the values are drawn from a variable rather then
directly from the function call.
}

\examples{
\donttest{
getHoldings.SMIF(FALSE, ipkey="The IP Key", pw="The Server Password")
}
}
\seealso{
Other data retrieval functions: \code{\link{getSectorETF}},
  \code{\link{getStockInfo}}, \code{\link{getSymbols.SMIF}}
}
\keyword{data}
\keyword{misc}
