% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectorETF.R
\name{getSectorETF.sector}
\alias{getSectorETF.sector}
\title{GetSectorETF.sector}
\usage{
getSectorETF.sector(sector, verbose = getOption("verbose", FALSE))
}
\description{
GetSectorETF.sector
}
