% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectorWeights.R
\name{getSectorWeights}
\alias{getSectorWeights}
\alias{getSectorWeightings}
\alias{getSectorWeights.SPY}
\alias{getSectorWeightings.SPY}
\alias{getSectorWeights.SPY}
\title{Load sector weightings for indexes}
\usage{
getSectorWeights(index = "SPY")

getSectorWeights.SPY()
}
\arguments{
\item{index}{Character; string specifying the name of the index. One of \code{"SPY"}.}
}
\value{
a data.frame containing \code{index}'s sector weightings (11 columns)
}
\description{
Function to scrape, clean, and format sector weightings for a given index or ETF.
}
\details{
\code{getSectorWeights} will try to coerce the value of the \code{index} argument into a recognizable ticker using
\code{\link{cleanIndex}} before it attempts to access sector weightings. Data will be loaded from different sources
depending on input arguments:

For S&P 500 data, data will be retreived from the official
\href{"http://us.spindices.com/documents/additional-material/500-sector-representation.xlsx?force_download=true"}{S&P Dow Jones website}
}
\examples{
getSectorWeights("S&P 500")
getSectorWeights.SPY()
}
\seealso{
\code{\link{smif.package}}\code{\link{cleanIndex}}
}
\author{
Alec Kulakowski, \email{alecthekulak@gmail.com}
}
\keyword{data}
\keyword{misc}
