% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlides.R
\name{getSlides}
\alias{getSlides}
\title{Get data for SMIF risk slides}
\usage{
getSlides(ticker, sector = getStockInfo.sector(ticker = ticker),
  use.rfr = TRUE)
}
\arguments{
\item{ticker}{Character; the ticker for the stock}

\item{sector}{Character; the sector of the \code{ticker}. If ommited, will use \code{getStockInfo.sector}.}

\item{use.rfr}{Logical; whether to use the risk-free rate in the beta calculations. Defaults to \code{TRUE}.}
}
\value{
Data.frame; risk slide table with filled in values.
}
\description{
Produces the data currently required for pitch risk slides. Returns the default risk slide table with
filled-in values (except for implied volatility, which is omitted completely).
}
