% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockInfo.R
\name{getStockInfo}
\alias{getStockInfo}
\alias{getStockInfo.Sector}
\alias{getStockInfo.Name}
\alias{getStockInfo.Last.Trade}
\alias{getStockInfo.Industry}
\alias{getStockInfo.Exchange}
\alias{getStockInfo.IPO.Year}
\alias{getStockInfo.Market.Cap}
\alias{getStockInfo.Name}
\alias{getStockInfo.Sector}
\alias{getStockInfo.Exchange}
\alias{getStockInfo.Last.Trade}
\alias{getStockInfo.Industry}
\alias{getStockInfo.IPO.Year}
\alias{getStockInfo.Market.Cap}
\title{Retreives info for a stock}
\usage{
getStockInfo(ticker, numMarketCap = TRUE, auto.assign = FALSE,
  env = .GlobalEnv)

getStockInfo.Name(ticker)

getStockInfo.Sector(ticker)

getStockInfo.Exchange(ticker)

getStockInfo.Last.Trade(ticker)

getStockInfo.Industry(ticker)

getStockInfo.IPO.Year(ticker)

getStockInfo.Market.Cap(ticker, numMarketCap = TRUE)
}
\arguments{
\item{ticker}{Character; the ticker for the stock you need info about}

\item{numMarketCap}{Logical; should the Market.Cap element be returned cleaned. If \code{TRUE}:
Market.Cap returns as a numeric. If \code{FALSE}: Market.Cap returns as a character string.
Defaults to \code{TRUE}}

\item{auto.assign}{Logical; should the results be assigned to \code{env}. If \code{FALSE} returns results.
Defaults to \code{FALSE}}

\item{env}{Environment; where to auto.assign objects. Setting \code{env=NULL} is equal to
\code{auto.assign=FALSE}. Defaults to \code{.GlobalEnv}.}
}
\value{
\code{getStockInfo} returns a list containing the elements:

  \item{Name}{(character) name of the company}
  \item{Last.Trade}{(numeric) price of the last trade for the company}
  \item{Sector}{(character) sector of the company}
  \item{Industry}{(character) descriptive industry name}
  \item{IPO.Year}{(integer) year of the company's IPO}
  \item{Exchange}{(character) exchange that the company currently trades on. One of: \code{"AMEX"},
       \code{"NASDAQ"}, \code{NYSE}.}
  \item{Market.Cap}{(numeric or character) current market capitalization for the company}
}
\description{
Functions to retrieve specific info for a given ticker. Based off of the \code{TTR} function \code{stockSymbols}
it compensates for different data types and difficulty in accessing specific ticker data.

Loads the last trade price for the given ticker
}
\details{
Data retrieval depends on the \code{TTR} package, which retrieves its data from \href{http://www.nasdaq.com/}{NASDAQ}
}
\note{
Values for \code{Sector} may not be identical to sectors listed in \code{\link{getSectorWeights}}.
Values for \code{Sector}, \code{IPO.Year}, and \code{Market.Cap} are occasionally missing and will default to \emph{NA}.
}
\examples{
getStockInfo("NVDA")
getStockInfo.Name("NVDA")
\donttest{getStockInfo.Sector("NVDA") }
\donttest{getStockInfo.Exchange("NVDA") }
\donttest{getStockInfo.Last.Trade("NVDA") }
\donttest{getStockInfo.Industry("NVDA") }
\donttest{getStockInfo.IPO.Year("NVDA") }
\donttest{getStockInfo.Market.Cap("NVDA", numMarketCap=FALSE) }
}
\author{
Alec Kulakowski, \email{alecthekulak@gmail.com}
}
\keyword{data}
\keyword{misc}
