% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.SMIF.R
\name{getSymbols.SMIF}
\alias{getSymbols.SMIF}
\title{Get stock data (optimized)}
\usage{
getSymbols.SMIF(tickers, full = F, try.av = F, ret = T)
}
\arguments{
\item{tickers}{Character; a string or a vector of strings specifying the data which is required}

\item{full}{Logical; should function return "full" data? If \code{TRUE}: return data will be a
\code{list} of \code{OHLC} objects. If \code{FALSE}: return data will be an \code{xts} of
Adjusted/Close price data for each ticker. Defaults to \code{FALSE}}

\item{try.av}{Logical; should program attempt to get data from Alpha Vantage? Defaults to \code{FALSE}
(AV servers are currently slow compared to Google's)}

\item{ret}{Logical; should results be returned to user. If \code{FALSE} returns \code{tickers} to
user. Defaults to \code{TRUE}}
}
\value{
Either a \code{list} of \code{OHLC} objects (if \code{isTRUE(full)}) or a \code{xts} of
Adjusted price data for each ticker (if \code{!isTRUE(full)})
}
\description{
A custom wrapper for getSymbol calls. Switches between different sources depending on server
status, and ticker classification. Negates need for local storage of price data between calls.
Optimized to increase speed of subsequent calls dramatically.
}
\details{
Uses \code{tryCatch} and global options to determine and store current server status for premium
data provider Alpha Vantage. Retrieves data based on status of AV's server (if down, data is
retrieved from Google). Retrieved data is stored in a private environment \emph{and} returned.
This makes subsequent calls to the function with the same or similar arguments practically
instantaneous. \code{full} data currently unavailable for \code{tickers} which contain "RFR".
}
\examples{
\dontrun{
getSymbols.SMIF(c("NVDA", "SPY", "RFR"), full = FALSE) }
getSymbols.SMIF("SPY", ret = FALSE)
}
\seealso{
\code{\link{.getTo}}, \code{\link{.getFrom}}: used to determine date boundries for data
retrieval

Other data retrieval functions: \code{\link{getHoldings.SMIF}},
  \code{\link{getSectorETF}}, \code{\link{getStockInfo}}
}
