% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\name{indexIdentifier}
\alias{indexIdentifier}
\title{Valid identifiers for common indexes}
\format{Character; vector of index identifiers as strings. (Current length = 4)}
\source{
\link{indexIdentifier}
}
\usage{
indexIdentifier()
}
\description{
A dataset containing index identifiers in a uniform manner. This dataset is currently a stub.
More values will be added as \code{cleanIndex} functionality is expanded. All elements of
\code{indexIdentifiers} currently work as explicit or rough inputs to \code{cleanIndex}.
}
\examples{
"SPY" \%in\% indexIdentifier()
}
\keyword{internal}
