% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smif.package.R
\docType{package}
\name{smif.package-package}
\alias{smif.package}
\alias{smif.package-package}
\title{\code{smif.package} allows members of the SMIF Risk and Asset Allocation teams to store commonly used functions and
datasets for use accross multiple devices and within RShiny projects. The package will be hosted on github, and can 
be downloaded with the following code:}
\description{
\code{devtools::install_github('alec25/smif.package')}
}
\details{
The current functions held by \code{smif.package} are as follows: \code{\link{getConstituents}}, 
\code{\link{getConstituents.simple}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alec25/smif.package}
  \item Report bugs at \url{https://github.com/alec25/smif.package/issues}
}

}
\author{
\strong{Maintainer}: Alec Kulakowski \email{alecthekulak@gmail.com}

}
