% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\name{cleanIndex}
\alias{cleanIndex}
\title{Clean index inputs}
\source{
\link{indexIdentifier}
}
\usage{
cleanIndex(index, stops = TRUE)
}
\arguments{
\item{index}{Character; string that must be coerced into an index identifier}

\item{stops}{Logical; whether the function should stop and throw an error message if the ticker
cannot be coerced}
}
\value{
Character; string representation of the given index
}
\description{
This function uses input parsing and replacement to coerce the character vector input
into the index that the user was referencing. Primary useage is found in other
\code{smif.package} functions which require uniform index identifiers.
}
\details{
\code{cleanIndex} will attempt to remove all non alphanumeric characters from \code{index} and
will coerce all alphabetic characters to lowercase. It then compares them to possible formatted spellings
of various indexes from \code{indexIdentifier} until a match is found. If no match is found, returns an error.
}
\examples{
cleanIndex("S&P 500")
}
\seealso{
\code{\link{indexIdentifier}}: internal dataset used to identify trivial cases

Other data cleaning functions: \code{\link{cleanAccount}},
  \code{\link{cleanIP}}, \code{\link{cleanSector}}
}
\concept{
clean index
}
