% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstituents.R
\name{getConstituents}
\alias{getConstituents}
\title{Load Constituent Data for Indexes}
\usage{
getConstituents(index, env = .GlobalEnv, simple = TRUE, src = "default",
  auto.assign = TRUE)
}
\arguments{
\item{index}{character specifying the name of the index}

\item{env}{where to create objects. Setting env=NULL is equal to auto.assign=FALSE. Defaults to .GlobalEnv}

\item{simple}{should results include non-ticker data. If \code{TRUE}, returns character vector of \code{index}'s 
constituents. If \code{FALSE} returns an. Defaults to \code{TRUE}}

\item{src}{character string specifying the source to retrieve the data from. As of 0.1-1, 
the only source available is "default" which changes depending on the selected \code{index}. Defaults to "default"}

\item{auto.assign}{should results be loaded to \code{env}. If \code{FALSE} return results. Defaults to \code{TRUE}}
}
\value{
if \code{auto.assign==FALSE}:  a character vector of \code{index}'s constituents

if \code{auto.assign==TRUE}:  the name of the variable holding the data within the specified environment
}
\description{
Functions to scrape and clean index constituent data from multiple sources for a given index. 
Current \code{src} methods available are: default.
Data is loaded silently without user assignment by default.
}
\examples{
## Not run: 
getConstituents(index='S&P500')
## End(Not run)
}
