% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHoldings.SMIF.R
\name{getHoldings.SMIF}
\alias{getHoldings.SMIF}
\alias{getHoldings.smif}
\title{Loads the current SMIF holdings}
\usage{
getHoldings.SMIF(auto.assign = TRUE,
  ipkey = readline("What is our favorite bank: "),
  serverPW = readline("Enter the server password: "), env = .GlobalEnv,
  varName = "SMIF.Holdings")
}
\arguments{
\item{auto.assign}{Logical; should results be assigned to variable in the environment specified by \code{env}.
Defaults to \code{TRUE}}

\item{ipkey}{Character; string answer to decryption question}

\item{serverPW}{Character; string of the SQL server's password}

\item{env}{Environment; specifies the environment to assign the results to. If this is not a valid \code{environment}
then auto.assign will evaluate as \code{FALSE}. Defaults to \code{.GlobalEnv}}

\item{varName}{Character; string for the name of the variable being assigned. Defaults to \code{SMIF.Holdings}}
}
\value{
a data.frame with the following columns:
\item{Ticker}{The ticker of the given holding}
\item{Shares}{The number of shares we currently own}
}
\description{
Fetches the current holdings for the SMIF from the server. Data returned will be sufficient to conduct
portfolio optimization on its own. Additional analysis on the information is possible in conjunction with
the other \code{smif.package} functions.
}
\details{
For compliance purposes, in order to connect to the server the connection info is not stored in plaintext.
The password is inputted directly by the user at runtime, and the server's IP address is encrypted. Decryption
occurs at runtime via answering a question that members are aware of but the general public does not know.
Function uses specific error handling and custom error messages.
}
\note{
Hello! This is a note
}
\section{Warning}{

Due to security settings on our SQL server, for certain R sessions the DBI connections are severed improperly.
When DBI connections are left open too long they expire and subsequent DBI connections to the server are
impossible until a new session is started. For this reason it is reccomended the results be stored in an
environment variable as soon as they are downloaded, and the values are drawn from a variable rather then
directly from the function call.
}

\examples{
\donttest{
getHoldings.SMIF(FALSE, ipkey="The IP Key", serverPW="The Server Password")
}
}
\author{
Alec Kulakowski, \email{alecthekulak@gmail.com}
}
\keyword{data}
\keyword{misc}
