% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectorETF.R
\name{getSectorETF}
\alias{getSectorETF}
\alias{getSectorETF.sector}
\title{Identify ETFs that correspond to a given sector}
\usage{
getSectorETF(ticker, sector)

getSectorETF.sector(sector)
}
\arguments{
\item{ticker}{Character; string of a stock's ticker}

\item{sector}{Character; string of a sector}
}
\value{
Character; string of sector ETF's ticker
}
\description{
Uses \code{getStockInfo.sector} and/or \code{cleanSector} to
retrieve uniform sector name for input. One of either
\code{ticker} or \code{sector} must be supplied.
}
\note{
\code{getSectorETF} is a wrapper of various sub-functions.
If both \code{ticker} and \code{sector} are supplied,
\code{sector} will take priority.
}
\examples{
getSectorETF(ticker = "NVDA")
getSectorETF(sector = "Consumer Non-Durables")
getSectorETF.sector("Telecom")
}
\seealso{
\code{\link{getStockInfo.sector}}: to retrieve sector
that corresponds to input ticker. \code{\link{cleanSector}}: to
coerce sector into uniform name.

Other data retrieval functions: \code{\link{getHoldings.SMIF}},
  \code{\link{getStockInfo}}, \code{\link{getSymbols.SMIF}}
}
\keyword{internal}
