% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHoldings.SMIF.R
\name{getServerData}
\alias{getServerData}
\title{Retrieves data from the SMIF server}
\usage{
getServerData(what, ...)
}
\arguments{
\item{what}{Character; string indicating what to retrieve. Valid entries
include 'holdings', 'cash balance'.}

\item{\dots}{additional parameters}
}
\value{
Data.frame; current holdings for the SMIF (if \code{what=="holdings"})
xts; cash balance history of the SMIF (if \code{what=="cash"})
}
\description{
Retrieves data from the SMIF server
}
