% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlides.R
\name{getSlides}
\alias{getSlides}
\title{Get data for SMIF risk slides}
\usage{
getSlides(ticker, sector = getStockInfo.sector(ticker = ticker),
  use.rfr = TRUE)
}
\arguments{
\item{ticker}{Character; the ticker for the stock}

\item{sector}{Character; the sector of the \code{ticker}. If ommited, will use \code{getStockInfo.sector}.}

\item{use.rfr}{Logical; whether to use the risk-free rate in the beta calculations. Defaults to \code{TRUE}.}
}
\value{
Data.frame; risk slide table with filled in values.
}
\description{
Produces the data currently required for pitch risk slides. Returns the default risk slide table (as a
data.frame) with filled-in values (except for implied volatility, which is omitted completely).
}
\details{
Provides data for the stock's volatility, sector beta, and market beta, based on 6 month data, and 12
month historical data. Betas are calculated manually and can incorporate the risk free rate if specified.
If sector cannot be identified, sector beta and market beta will be the same.
}
\examples{
\dontrun{
getSlides("NVDA", sector="Technology", use.rfr = TRUE)
}
}
\seealso{
\code{\link{getStockInfo.sector}}: identifies sector of ticker if not specified.
\code{\link{getSectorETF.sector}}: identifies sector ETF to be used for sector beta calculations.

Other data processing functions: \code{\link{getRiskMetrics}}
}
