% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockInfo.R
\name{getStockInfo}
\alias{getStockInfo}
\alias{getStockInfo.sector}
\alias{getStockInfo.name}
\alias{getStockInfo.last}
\alias{getStockInfo.industry}
\alias{getStockInfo.exchange}
\alias{getStockInfo.IPO.year}
\alias{getStockInfo.market.cap}
\alias{getStockInfo.name}
\alias{getStockInfo.sector}
\alias{getStockInfo.exchange}
\alias{getStockInfo.last}
\alias{getStockInfo.industry}
\alias{getStockInfo.IPO.year}
\alias{getStockInfo.market.cap}
\title{Retreives info for a stock}
\usage{
getStockInfo(ticker, numMarketCap = TRUE, auto.assign = FALSE,
  env = .GlobalEnv)

getStockInfo.name(ticker)

getStockInfo.sector(ticker)

getStockInfo.exchange(ticker)

getStockInfo.last(ticker)

getStockInfo.industry(ticker)

getStockInfo.IPO.year(ticker)

getStockInfo.market.cap(ticker, numMarketCap = TRUE)
}
\arguments{
\item{ticker}{Character; the ticker for the stock you need info about}

\item{numMarketCap}{Logical; should the Market.Cap element be returned cleaned. If \code{TRUE}:
Market.Cap returns as a numeric. If \code{FALSE}: Market.Cap returns as a character string.
Defaults to \code{TRUE}}

\item{auto.assign}{Logical; should the results be assigned to \code{env}. If \code{FALSE} returns results.
Defaults to \code{FALSE}}

\item{env}{Environment; where to auto.assign objects. Setting \code{env=NULL} is equal to
\code{auto.assign=FALSE}. Defaults to \code{.GlobalEnv}.}
}
\value{
\code{getStockInfo} returns a list containing the elements:

  \item{Name}{Character; name of the company}
  \item{Last.Trade}{Numeric; price of the last trade for the company}
  \item{Sector}{Character; sector of the company}
  \item{Industry}{Character; descriptive industry name}
  \item{IPO.Year}{Integer; year of the company's IPO}
  \item{Exchange}{Character; exchange that the company currently trades on. One of: \code{"AMEX"},
       \code{"NASDAQ"}, \code{NYSE}.}
  \item{Market.Cap}{Numeric (or Character); current market capitalization for the company}
}
\description{
Function family to retrieve specific info for a given ticker.
}
\details{
Based off of the \code{TTR} function \code{stockSymbols} it compensates for different data types and difficulty
in accessing specific ticker data. Data retrieval depends on the \code{TTR} package, which retrieves its data
from \href{http://www.nasdaq.com/}{NASDAQ}
}
\note{
Values for \code{Sector} may not be identical to sectors listed in \code{\link{getSectorWeights}}.
Values for \code{Sector}, \code{IPO.Year}, and \code{Market.Cap} are occasionally missing and will default to \emph{NA}.
}
\examples{
getStockInfo("NVDA")
getStockInfo.name("NVDA")
\donttest{getStockInfo.sector("NVDA") }
\donttest{getStockInfo.exchange("NVDA") }
\donttest{getStockInfo.last("NVDA") }
\donttest{getStockInfo.industry("NVDA") }
\donttest{getStockInfo.IPO.year("NVDA") }
\donttest{getStockInfo.market.cap("NVDA", numMarketCap=FALSE) }
}
\seealso{
Other data retrieval functions: \code{\link{getSectorETF}}
}
\concept{
getStockInfo.last.trade getStockInfo.price getStockInfo.last.price

getStockInfo.IPO

getStockInfo.mcap
}
\keyword{data}
\keyword{misc}
