% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockInfo.R
\name{getStockInfo}
\alias{getStockInfo}
\alias{getStockInfo.sector}
\alias{getStockInfo.name}
\alias{getStockInfo.last}
\alias{getStockInfo.industry}
\alias{getStockInfo.exchange}
\alias{getStockInfo.IPO.year}
\alias{getStockInfo.market.cap}
\alias{getStockInfo.last.trade}
\alias{getStockInfo.price}
\alias{getStockInfo.last.price}
\alias{getStockInfo.mcap}
\alias{getStockInfo.IPO}
\alias{getStockInfo.name}
\alias{getStockInfo.sector}
\alias{getStockInfo.exchange}
\alias{getStockInfo.last}
\alias{getStockInfo.industry}
\alias{getStockInfo.IPO.year}
\alias{getStockInfo.market.cap}
\title{Retreives info for a stock}
\usage{
getStockInfo(ticker, clean.mcap = TRUE, clean.sector = TRUE,
  auto.assign = FALSE, env = .GlobalEnv)

getStockInfo.name(ticker)

getStockInfo.sector(ticker, clean.sector = TRUE)

getStockInfo.exchange(ticker)

getStockInfo.last(ticker)

getStockInfo.industry(ticker)

getStockInfo.IPO.year(ticker)

getStockInfo.market.cap(ticker, clean.mcap = TRUE)
}
\arguments{
\item{ticker}{Character; the ticker for the stock you need info about}

\item{clean.mcap}{Logical; should the Market.Cap element be returned cleaned. If \code{TRUE}:
Market.Cap returns as a numeric. If \code{FALSE}: Market.Cap returns as a character string.
Defaults to \code{TRUE}}

\item{clean.sector}{Logical; should the Sector element be returned cleaned. Sector cleaning is done using
\code{\link{cleanSector}}. Defaults to \code{TRUE}}

\item{auto.assign}{Logical; should the results be assigned to \code{env}. If \code{FALSE} returns results.
Defaults to \code{FALSE}}

\item{env}{Environment; where to auto.assign objects. Setting \code{env=NULL} is equal to
\code{auto.assign=FALSE}. Defaults to \code{.GlobalEnv}.}
}
\value{
Returns a list containing the elements:

  \item{Name}{Character; name of the company}
  \item{Last.Trade}{Numeric; price of the last trade for the company}
  \item{Sector}{Character; sector of the company}
  \item{Industry}{Character; descriptive industry name}
  \item{IPO.Year}{Integer; year of the company's IPO}
  \item{Exchange}{Character; exchange that the company currently trades on. One of: \code{"AMEX"},
       \code{"NASDAQ"}, \code{NYSE}.}
  \item{Market.Cap}{Numeric (or Character); current market capitalization for the company}
}
\description{
Function family to retrieve specific info for a given ticker.
}
\details{
Based off of the \code{TTR} function \code{stockSymbols} it compensates for different data types and difficulty
in accessing specific ticker data. Data retrieval depends on the \code{TTR} package, which retrieves its data
from \href{http://www.nasdaq.com/}{NASDAQ}
}
\section{Functions}{
\itemize{
\item \code{getStockInfo.name}: Loads the company name for the given ticker (if available)

\item \code{getStockInfo.sector}: Loads the sector for the given ticker, allows changing the \code{clean.sector} variable

\item \code{getStockInfo.exchange}: Loads the exchange for the given ticker (if available)

\item \code{getStockInfo.last}: Loads the last trade price for the given ticker

\item \code{getStockInfo.industry}: Loads the last descriptive industry name for the given ticker

\item \code{getStockInfo.IPO.year}: Loads the IPO year for the given ticker (if available)

\item \code{getStockInfo.market.cap}: Loads the current market capitalization for the given ticker (if available), allows changing the \code{clean.mcap} variable
}}

\note{
Values for \code{Sector} may not be identical to sectors listed in \code{\link{getSectorWeights}}.
Values for \code{Sector}, \code{IPO.Year}, and \code{Market.Cap} are occasionally missing and will default to
\emph{NA}. \code{getStockInfo} caches the data for \code{stockSymbols} in a hidden global variable (\code{.ss})
to massively increase speed on subsequent uses of the function.
}
\examples{
getStockInfo("NVDA")
getStockInfo.name("NVDA")
\donttest{getStockInfo.sector("NVDA") }
\donttest{getStockInfo.exchange("NVDA") }
\donttest{getStockInfo.last("NVDA") }
\donttest{getStockInfo.industry("NVDA") }
\donttest{getStockInfo.IPO.year("NVDA") }
\donttest{getStockInfo.market.cap("NVDA", clean.mcap=FALSE) }
}
\seealso{
Other data retrieval functions: \code{\link{getSectorETF}}
}
\concept{
getStockInfo.last.trade getStockInfo.price getStockInfo.last.price
}
\keyword{data}
\keyword{internal}
\keyword{misc}
