% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIndex.R
\docType{data}
\name{indexIdentifier}
\alias{indexIdentifier}
\title{Valid identifiers for common indexes}
\format{A character vector of possible identifiers}
\usage{
indexIdentifier
}
\description{
A dataset containing index identifiers in a uniform manner
}
\keyword{internal}
