% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smif.package.R
\docType{package}
\name{smif.package-package}
\alias{smif.package}
\alias{smif.package-package}
\title{SMIF Function and Documentation Package}
\description{
The \code{smif.package} allows members of the SMIF Risk and Asset Allocation teams to store commonly used functions and
datasets for use accross multiple devices and within RShiny projects.
}
\details{
\tabular{ll}{
  Package: \tab smif.package \cr
  Type: \tab Package \cr
  Version: \tab 0.1-4 \cr
  Date: \tab 2017-10-04 \cr
  License: \tab GPL-3 \cr
  URL: \tab https://github.com/alec25/smif.package \cr
}

The package will be hosted on github, and the latest version can be downloaded with the following code: 
\code{devtools::install_github('alec25/smif.package')} 

The current master functions held by \code{smif.package} are as follows: 
\describe{
  \item{\code{\link{getConstituents}}}{Returns the constituents for a given index}
  \item{\code{\link{getSectorWeights}}}{Returns the sector weights for a given index}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alec25/smif.package}
  \item Report bugs at \url{https://github.com/alec25/smif.package/issues}
}

}
\author{
\strong{Maintainer}: Alec Kulakowski \email{alecthekulak@gmail.com}

}
