% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smif.package.R
\docType{package}
\name{smif.package}
\alias{smif.package}
\alias{smif.package-package}
\alias{smif.package-package}
\title{SMIF Function and Documentation Package}
\description{
The \code{smif.package} allows members of the SMIF Risk and Asset Allocation teams to store commonly used functions and
datasets for use accross multiple devices and within RShiny projects.
}
\details{
\tabular{ll}{
  Package: \tab smif.package \cr
  Type: \tab Package \cr
  Version: \tab 0.1-12 \cr
  Date: \tab 2017-10-20 \cr
  License: \tab GPL-3 + file LICENSE \cr
  Depends: \tab magrittr(>=1.5) \cr
  URL: \tab https://github.com/alec25/smif.package \cr
}

The package will be hosted on github, and the latest version can be downloaded with the following code:
\code{devtools::install_github('alec25/smif.package', dependencies=TRUE, upgrade_dependencies=TRUE)}
}
\section{Functions}{


The current master functions held by \code{smif.package} are as follows:
\describe{
  \item{\code{\link{getConstituents}}}{Returns the current constituents for a given index}
  \item{\code{\link{getSectorWeights}}}{Returns the sector weightings for a given index}
  \item{\code{\link{getStockInfo}}}{Returns info for a specified stock}
  \item{\code{\link{getHoldings.SMIF}}}{Returns info for a specified stock}
  \item{\code{\link{getSectorETF}}}{Returns the S&P ETF for a given sector}
  \item{\code{\link{getSlides}}}{Returns the figures needed for pitch risk slides}
}
The current datasets provided by \code{smif.package} are:
\describe{
  \item{\code{\link{smif_aa}}}{Assorted datasets with Asset Allocation applications}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alec25/smif.package}
  \item Report bugs at \url{https://github.com/alec25/smif.package/issues}
}

}
\author{
\strong{Maintainer}: Alec Kulakowski \email{alecthekulak@gmail.com}

}
