% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smif_aa-data.R
\docType{data}
\name{smif_aa}
\alias{smif_aa}
\title{Collection of data relevent to SMIF Asset Allocation}
\format{A list of various datasets:
\describe{
  \item{\code{sectors}}{Character vector of all possible sector names}
}}
\source{
Marco Gancitano (\url{mgancita@stevens.edu})
}
\usage{
data(smif_aa)
}
\description{
Currently only contains one sub-dataset. More will be added in
future. Contact package administrator to suggest an additoin.
}
\note{
The \code{sectors} dataset does not include any non-equity sectors, such as REITs, nor any
"miscellaneous" sector categorization such as "other."
}
\examples{
\donttest{
data(smif_aa)
smif_aa$sectors
}
}
\keyword{datasets}
